refactor: remove dead rebuild code
